Initial commit
This is basically a copy of old code from the DokuWiki release
Rincewind transformed into an action plugin. The output is still
unstyled and has known problems like redirecting to the media popup in
all cases (also from the new media manager) but in a first test
uploading multiple files in Internet Explorer 8 worked.